requirements.txt
blessed==1.20.0
getkey==0.6.5
inquirerpy==0.3.4
loguru==0.7.2
pfzy==0.3.4
prompt-toolkit==3.0.43
psycopg==3.1.15
python-dotenv==1.0.0
python-editor==1.0.4
readchar==4.0.5
simple-term-menu==1.6.3
six==1.16.0
tabulate==0.9.0
typing_extensions==4.9.0
wcwidth==0.2.12
